django-backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
pytest==8.3.3
pytest-django==4.9.0

# Authentication & HTTP
PyJWT==2.8.0
requests==2.31.0
cryptography==42.0.5

# Production dependencies
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.4.0
django-storages==1.14.4
boto3==1.35.8
Pillow==10.4.0
drf-spectacular==0.27.2
